Update the Makefile.
The Makefile now stop properly the docker compose and all its containers. Moreover, the down command now stop and remove all the containers, including the database. In consequences, the life cycle of the project is now to use docker up to build and run, and docker stop to stop the containers. Because the down command also remove the database, only use it when you want to reset TotemBionet to a clean state.